parametro departamento al mapa
Se le agrego el parametro de departamento al mapa
